move searchDropdown to be closer to parent component
