+++ src/sentry/static/sentry/app/components/smartSearchBar/searchDropdown.jsx
